Code Quality:
* Separate more TrainManager stuff.
