+++ openBVE/OpenBve/OldCode/TrainManager.cs
@@ -66,6 +66,7 @@ namespace OpenBve
 			internal double EqualizingReservoirCurrentPressure;
 			internal double EqualizingReservoirNormalPressure;
 			internal double EqualizingReservoirServiceRate;
+
 			internal double EqualizingReservoirEmergencyRate;
 			internal double EqualizingReservoirChargeRate;
 			internal double BrakePipeCurrentPressure;
@@ -2559,101 +2560,7 @@ namespace OpenBve
 				Train.Plugin.UpdatePower();
 				Train.Plugin.UpdateBrake();
 			}
-		}
-
-		// apply air brake handle
-		internal static void ApplyAirBrakeHandle(Train Train, int RelativeDirection)
-		{
-			if (Train.Cars[Train.DriverCar].Specs.BrakeType == CarBrakeType.AutomaticAirBrake)
-			{
-				if (RelativeDirection == -1)
-				{
-					if (Train.Specs.AirBrake.Handle.Driver == AirBrakeHandleState.Service)
-					{
-						ApplyAirBrakeHandle(Train, AirBrakeHandleState.Lap);
-					}
-					else
-					{
-						ApplyAirBrakeHandle(Train, AirBrakeHandleState.Release);
-					}
-				}
-				else if (RelativeDirection == 1)
-				{
-					if (Train.Specs.AirBrake.Handle.Driver == AirBrakeHandleState.Release)
-					{
-						ApplyAirBrakeHandle(Train, AirBrakeHandleState.Lap);
-					}
-					else
-					{
-						ApplyAirBrakeHandle(Train, AirBrakeHandleState.Service);
-					}
-				}
-				Game.AddBlackBoxEntry(Game.BlackBoxEventToken.None);
-			}
-		}
-		internal static void ApplyAirBrakeHandle(Train Train, AirBrakeHandleState State)
-		{
-			if (Train.Cars[Train.DriverCar].Specs.BrakeType == CarBrakeType.AutomaticAirBrake)
-			{
-				if (State != Train.Specs.AirBrake.Handle.Driver)
-				{
-					// sound when moved to service
-					if (State == AirBrakeHandleState.Service)
-					{
-						Sounds.SoundBuffer buffer = Train.Cars[Train.DriverCar].Sounds.Brake.Buffer;
-						if (buffer != null)
-						{
-							OpenBveApi.Math.Vector3 pos = Train.Cars[Train.DriverCar].Sounds.Brake.Position;
-							Sounds.PlaySound(buffer, 1.0, 1.0, pos, Train, Train.DriverCar, false);
-						}
-					}
-					// sound
-					if ((int)State < (int)Train.Specs.AirBrake.Handle.Driver)
-					{
-						// brake release
-						if ((int)State > 0)
-						{
-							// brake release (not min)
-							Sounds.SoundBuffer buffer = Train.Cars[Train.DriverCar].Sounds.BrakeHandleRelease.Buffer;
-							if (buffer != null)
-							{
-								OpenBveApi.Math.Vector3 pos = Train.Cars[Train.DriverCar].Sounds.BrakeHandleRelease.Position;
-								Sounds.PlaySound(buffer, 1.0, 1.0, pos, Train, Train.DriverCar, false);
-							}
-						}
-						else
-						{
-							// brake min
-							Sounds.SoundBuffer buffer = Train.Cars[Train.DriverCar].Sounds.BrakeHandleMin.Buffer;
-							if (buffer != null)
-							{
-								OpenBveApi.Math.Vector3 pos = Train.Cars[Train.DriverCar].Sounds.BrakeHandleMin.Position;
-								Sounds.PlaySound(buffer, 1.0, 1.0, pos, Train, Train.DriverCar, false);
-							}
-						}
-					}
-					else if ((int)State > (int)Train.Specs.AirBrake.Handle.Driver)
-					{
-						// brake
-						Sounds.SoundBuffer buffer = Train.Cars[Train.DriverCar].Sounds.BrakeHandleApply.Buffer;
-						if (buffer != null)
-						{
-							OpenBveApi.Math.Vector3 pos = Train.Cars[Train.DriverCar].Sounds.BrakeHandleApply.Position;
-							Sounds.PlaySound(buffer, 1.0, 1.0, pos, Train, Train.DriverCar, false);
-						}
-					}
-					// apply
-					Train.Specs.AirBrake.Handle.Driver = State;
-					Game.AddBlackBoxEntry(Game.BlackBoxEventToken.None);
-					// plugin
-					if (Train.Plugin != null)
-					{
-						Train.Plugin.UpdatePower();
-						Train.Plugin.UpdateBrake();
-					}
-				}
-			}
-		}
+		}	
 
 		// update train passengers
 		
@@ -3221,132 +3128,6 @@ namespace OpenBve
 			Train.Specs.CurrentAverageAcceleration *= invcarlen;
 		}
 
-
-		// un-derail train
-		internal static void UnderailTrains()
-		{
-			System.Threading.Tasks.Parallel.For(0, Trains.Length, i =>
-			{
-				UnderailTrain(Trains[i]);
-			});
-			//for (int i = 0; i < Trains.Length; i++) {
-			//	UnderailTrain(Trains[i]);
-			//}
-		}
-		internal static void UnderailTrain(Train Train)
-		{
-			for (int i = 0; i < Train.Cars.Length; i++)
-			{
-				Train.Cars[i].Specs.CurrentRollDueToTopplingAngle = 0.0;
-				Train.Cars[i].Derailed = false;
-			}
-		}
-
-		// jump train
-		internal static void JumpTrain(Train train, int stationIndex)
-		{
-			train.StationState = TrainStopState.Pending;
-			int stopIndex = Game.GetStopIndex(stationIndex, train.Cars.Length);
-			if (stopIndex >= 0)
-			{
-				if (train == PlayerTrain)
-				{
-					if (train.Plugin != null)
-					{
-						train.Plugin.BeginJump((OpenBveApi.Runtime.InitializationModes)Game.TrainStart);
-					}
-				}
-				for (int h = 0; h < train.Cars.Length; h++)
-				{
-					train.Cars[h].Specs.CurrentSpeed = 0.0;
-				}
-				double d = Game.Stations[stationIndex].Stops[stopIndex].TrackPosition - train.Cars[0].FrontAxle.Follower.TrackPosition + train.Cars[0].FrontAxlePosition - 0.5 * train.Cars[0].Length;
-				if (train == PlayerTrain)
-				{
-					TrackManager.SuppressSoundEvents = true;
-				}
-				while (d != 0.0)
-				{
-					double x;
-					if (Math.Abs(d) > 1.0)
-					{
-						x = (double)Math.Sign(d);
-					}
-					else
-					{
-						x = d;
-					}
-					for (int h = 0; h < train.Cars.Length; h++)
-					{
-						TrainManager.MoveCar(train, h, x, 0.0);
-					}
-					if (Math.Abs(d) >= 1.0)
-					{
-						d -= x;
-					}
-					else
-					{
-						break;
-					}
-				}
-				if (train == PlayerTrain)
-				{
-					TrainManager.UnderailTrains();
-					TrackManager.SuppressSoundEvents = false;
-				}
-				if (train.Specs.CurrentEmergencyBrake.Driver)
-				{
-					TrainManager.ApplyNotch(train, 0, false, 0, true);
-				}
-				else
-				{
-					TrainManager.ApplyNotch(train, 0, false, train.Specs.MaximumBrakeNotch, false);
-					TrainManager.ApplyAirBrakeHandle(train, TrainManager.AirBrakeHandleState.Service);
-				}
-				if (Game.Sections.Length > 0)
-				{
-					Game.UpdateSection(Game.Sections.Length - 1);
-				}
-				if (train == PlayerTrain)
-				{
-					if (Game.CurrentScore.ArrivalStation <= stationIndex)
-					{
-						Game.CurrentScore.ArrivalStation = stationIndex + 1;
-					}
-				}
-				if (train == PlayerTrain)
-				{
-					if (Game.Stations[stationIndex].ArrivalTime >= 0.0)
-					{
-						Game.SecondsSinceMidnight = Game.Stations[stationIndex].ArrivalTime;
-					}
-					else if (Game.Stations[stationIndex].DepartureTime >= 0.0)
-					{
-						Game.SecondsSinceMidnight = Game.Stations[stationIndex].DepartureTime - Game.Stations[stationIndex].StopTime;
-					}
-				}
-				for (int i = 0; i < train.Cars.Length; i++)
-				{
-					train.Cars[i].Specs.AnticipatedLeftDoorsOpened = Game.Stations[stationIndex].OpenLeftDoors;
-					train.Cars[i].Specs.AnticipatedRightDoorsOpened = Game.Stations[stationIndex].OpenRightDoors;
-				}
-				if (train == PlayerTrain)
-				{
-					Game.CurrentScore.DepartureStation = stationIndex;
-					Game.CurrentInterface = Game.InterfaceType.Normal;
-					Game.Messages = new Game.Message[] { };
-				}
-				ObjectManager.UpdateAnimatedWorldObjects(0.0, true);
-				TrainManager.UpdateTrainObjects(0.0, true);
-				if (train == PlayerTrain)
-				{
-					if (train.Plugin != null)
-					{
-						train.Plugin.EndJump();
-					}
-				}
-			}
-		}
 
 		// update train physics and controls
 		private static void UpdateTrainPhysicsAndControls(Train Train, double TimeElapsed)
