+++ openBVE/OpenBve/OpenBve.csproj
@@ -130,6 +130,8 @@
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
     <Compile Include="Simulation\TrainManager\Doors.cs" />
+    <Compile Include="Simulation\TrainManager\Functions.cs" />
+    <Compile Include="Simulation\TrainManager\Train\BrakeSystem.AirBrake.cs" />
     <Compile Include="Simulation\TrainManager\Train\BrakeSystem.cs" />
     <Compile Include="Simulation\TrainManager\Train\Handles.cs" />
     <Compile Include="Simulation\TrainManager\Passengers.cs" />
